requirements.txt
boto3
numpy==1.26.4
scipy==1.11.4
faiss-cpu==1.7.4
langchain
langchain-community
langchain-aws
langchain-text-splitters
unstructured
unstructured[pdf]
